requirements.txt
flask==2.3.3
requests==2.31.0
tensorflow-cpu==2.15.0
numpy==1.26.4
jieba==0.42.1